+++ CHANGELOG.rst
@@ -22,6 +22,8 @@ Changed
 - Bug fix in ``simbad.lattice.latticesearch`` for duplicate entries from alternative unit cells
 - Bug fix for logging and error message handling prior to logger initialisation
 - Bug fix in ``simbad.util.pdb_util`` variable name 
+- ``ccp4_root`` function changed to ``CCP4RootDirectory`` class
+- ``ccp4_version`` function changed to ``CCP4Version`` class and version extracted from official CCP4 release file
 
 0.1.0
 -----